fix: restore constants.ts file with runtime values to resolve circular dependency
- Create constants.ts with defaultVideoAppCategories, MeetLocationType, MSTeamsLocationType
- Update imports in utils.ts, googlecalendar, office365calendar, and locations.ts
- This separates runtime values from type declarations to fix module resolution errors
- Revert routing-forms lint fixes that don't solve actual circular dependencies
